source-mysql: Split out query_events.go
The MySQL query event / DDL handling logic grew organically over
time and at this point it's become pretty large, enough so that
we should really split it out from the core replication stream
and transcoding-DML-into-documents logic in replication.go
